requirements.txt
faster-whisper>=1.0.0
pyannote.audio>=3.1
gradio>=4.0
python-docx>=1.0
torch>=2.0